refactor(data-events): add abstract Connector class for shared methods (#3015)
* refactor: add abstract Connector class for shared methods

* chore: remove unused imports
